Improve random spin generation
Sometimes the random number of spins generated is the same as the previous one. When this happens for a clock that must display the same angles for the next digit, the hands don't move at all. There is now a lesser chance of this happening as we look for a random number that is different from the previous one.